Move get all vcs
